+++ src/components/Sing/CharacterMenuButton/MenuButton.vue
@@ -148,10 +148,9 @@ import { computed, ref } from "vue";
 import { debounce } from "quasar";
 import SelectedCharacter from "./SelectedCharacter.vue";
 import { useStore } from "@/store";
-import { base64ImageToUri } from "@/helpers/base64Helper";
-import { SpeakerId, StyleId, EngineId } from "@/type/preload";
+import { SpeakerId, StyleId } from "@/type/preload";
 import { getStyleDescription } from "@/sing/viewHelper";
-import { asyncComputed } from "@/composables/asyncComputed";
+import { useEngineIcons } from "@/composables/useEngineIcons";
 
 const store = useStore();
 const uiLocked = computed(() => store.getters.UI_LOCKED);
@@ -241,18 +240,7 @@ const selectedStyleId = computed(
 // 複数エンジン
 const isMultipleEngine = computed(() => store.state.engineIds.length > 1);
 
-const engineIcons = asyncComputed(
-  () => store.state.engineManifests,
-  {} as Record<EngineId, string>,
-  async (engineManifests) => {
-    const engineIcons: Record<EngineId, string> = {};
-    for (const [engineId, manifest] of Object.entries(engineManifests)) {
-      engineIcons[EngineId(engineId)] = await base64ImageToUri(manifest.icon);
-    }
-
-    return engineIcons;
-  },
-);
+const engineIcons = useEngineIcons(() => store.state.engineManifests);
 </script>
 
 <style scoped lang="scss">
